Move hotReloader to dev-server
